get rid of services package - split it into rpc and caller
