+++ internal/caller/fieldwalker.go
@@ -1,4 +1,4 @@
-package services
+package caller
 
 import "github.com/jhump/protoreflect/desc"
