+++ internal/rpc/connectionoptions.go
@@ -1,4 +1,4 @@
-package services
+package rpc
 
 import (
 	"errors"
