+++ internal/rpc/grpcconnfactory.go
@@ -1,4 +1,4 @@
-package services
+package rpc
 
 import (
 	"log"
